Do not create symlink to pycsw.cfg if it already exists
In `my_init` scripts, do not create pycsw.cfg symlink if it already exists.
This was causing issues with restarting docker, as the scripts would
attempt to create the symlink and would fail if it existed, exiting the
script with an error and causing the container not to start unless it
was removed and re-run.