Adds better attribution
The Access tab got a small refactor. Each entry found within an ISO
defining CI_Responsible_Party should have a row in the table for Access.
The ISOTC211/19115 Roles are mapped to human readable versions and the
contents for the responsible party are rendered. URLs are mapped to HTML
Anchors. Emails are mapped to Anchors with `mailto` defined as the email
address.

I also added some code that cleans up edge cases where required tags
aren't filled in like names for URLs.