Adds feedback
This commit adds two new schema items: smtp.port (not sure why we didn't
have it before). As well as feedback.recipients which is a space
delimited list of email addresses that will receive feedback.

This commit adds a new controller called FeedbackController which simply
renders the "Submit Feedback" form and passes the template the data.

This commit adds a new library for mailing based on Flask-Mail. In my
opinion, Flask-Mail deals with far more edge cases for email than the
existing CKAN mailer.

To disable feedback leave the recipients entry blank.

There's also a template for the body of the feedback emails.