Fixes config bug
When a plugin extends the update_config_schema method, these apply to
configration schema definitions for the UI's Admin Controller, not the
application as a whole. So when ckan.ioos_theme.pycsw_config was defined
in here, administrator's couldn't save the config form page because
there wasn't a form for the config file.