Add IOOS DCAT profile, adding CF standard names
Adds an IOOS DCAT profile option, which adds any CF standard names as
an array of text literals to the (currently) proposed variableMeasured
schema.org term.  Also adds `ckanext.ioos_theme.profiles.IOOSDCATProfile` to
the list of entrypoints in setup.py so that it can be specified as in the
CKAN config.